requirements/test.txt
-r base.txt
-r test-acceptance.txt

coverage==5.1
ddt==1.0.0
mock==1.3.0
pep8==1.5.7
pylint<3.0
pytest-django

django-pyfs==2.0
edx-i18n-tools
git+https://github.com/edx/xblock-sdk.git@331e50bed2a3c9014d998cc4428e92bf658baf8d#egg=xblock-sdk
